backend/requirements.txt
fastapi
uvicorn
python-socketio
motor
python-dotenv
pydantic